Src/requirements.txt
absl-py==0.9.0
altgraph==0.17
astor==0.8.1
backcall==0.1.0
bleach==3.1.0
certifi==2018.8.24
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.2.3
future==0.18.2
gast==0.3.3
grpcio==1.27.2
h5py==2.10.0
idna==2.9
imageio==2.8.0
ipykernel==4.10.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
jedi==0.12.1
Jinja2==2.11.1
joblib==0.14.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==5.2.0
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.0
matplotlib==3.0.3
mistune==0.8.3
nbconvert==5.5.0
nbformat==5.0.4
networkx==2.4
notebook==5.6.0
numpy==1.15.4
opencv-python==3.4.3.18
pandocfilters==1.4.2
parso==0.6.2
pefile==2019.4.18
pickleshare==0.7.4
Pillow==7.0.0
pipreqs==0.4.10
prometheus-client==0.7.1
prompt-toolkit==1.0.15
protobuf==3.11.3
Pygments==2.6.1
PyInstaller==3.6
pyparsing==2.4.6
pypiwin32==223
pyserial==3.4
PySide2==5.14.1
python-dateutil==2.8.1
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.4
PyYAML==5.3.1
pyzmq==17.1.2
qtconsole==4.7.1
QtPy==1.9.0
requests==2.23.0
scikit-image==0.15.0
scikit-learn==0.22.2.post1
scipy==1.2.1
Send2Trash==1.5.0
shiboken2==5.14.1
simplegeneric==0.8.1
six==1.14.0
sklearn==0.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.4
Theano==1.0.4
tornado==5.1.1
traitlets==4.3.2
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.4.1
win-unicode-console==0.5
wincertstore==0.2
yarg==0.1.9
